fix(build): cast the enum fallback in the packed-token malloc; GCC -Werror=extra rejects an enum vs size_t conditional
Both Linux CI legs (ubuntu-latest and ubuntu-24.04-arm, gcc/g++) failed on
pass_semantic_edges.cpp:1273 after the 2026-09-20 sync landed:
  error: enumerated and non-enumerated type in conditional expression
Same class as 1ac77fdc. Clang accepts it, which is why the macOS suite and
the review round were green.

+++ src/pipeline/pass_semantic_edges.cpp
@@ -1270,7 +1270,7 @@ static void phase2_tokenize(const cbm_gbuf_node_t **node_ptrs, cbm_gbuf_t *gbuf,
     for (int w = 0; w < worker_count; w++) {
         total += bufs[w].count;
     }
-    char **packed = (char **)malloc((total ? total : SKIP_ONE) * sizeof(char *));
+    char **packed = (char **)malloc((total ? total : (size_t)SKIP_ONE) * sizeof(char *));
     if (packed) {
         size_t base = 0;
         for (int w = 0; w < worker_count; w++) {
